+++ rts/include/stg/MiscClosures.h
@@ -532,6 +532,7 @@ RTS_FUN_DECL(stg_unmaskAsyncExceptionszh);
 RTS_FUN_DECL(stg_myThreadIdzh);
 RTS_FUN_DECL(stg_labelThreadzh);
 RTS_FUN_DECL(stg_isCurrentThreadBoundzh);
+RTS_FUN_DECL(stg_threadLabelzh);
 RTS_FUN_DECL(stg_threadStatuszh);
 RTS_FUN_DECL(stg_listThreadszh);